requirements.txt
# requirements.txt
discord.py==2.5.2
python-dotenv==1.0.0
together==1.3.1
requests==2.31.0
